Introduce the 2.5 protocol upgrade
The protocol upgrade improves the network safety and fairness of reward distribution. After the
fork, SPoRA solution candidates have to be "packed" so that miners cannot explore patterns in data
to gain an extra advantage. Additionally, the protocol creates some entropy on every mining
attempt. Miners have to carry it on throughout the process, further strengthening the requirement
to keep data close to the computation power. Finally, the new version requires data to be split
in a particular way (compatible with the majority of the existing tools) to ensure all data chunks
are equally valuable for the miners and drastically simplifying the storage and indexing routines.

The amount of packing and unpacking the node performs does not exceed the configured packing rate.
By default it is 10 chunks per second matching the speed at which the pre-2.5 chunks start being
required in the packed form.

The new metrics: v2_index_data_size_by_packing, packing_duration_milliseconds,
validating_unpacked_spora, validating_packed_spora.

Co-authored-by: vird <virdvip@gmail.com>